// src/lib/api/types.d.ts
export interface operations {
// ...
  topicByArticleCount: {
    parameters: {
      query?: never
      header?: never
      path?: never
      cookie?: never
    }
    requestBody: {
      content: {
        "application/json": {
          /** @default 1 */
          page?: number
          /** @default 10 */
          perPage?: number
          /** @enum {string} */
          language: "id" | "en"
        }
      }
    }
    responses: {
      /** @description OK */
      200: {
        headers: Record<string, unknown>
        content: {
          "application/json": {
            id: string
            title: string
            slug: string
            language: string
            count: number
          }[]
        }
      }
    }
  }
// ...
  topicBySlug: {
    parameters: {
      query?: never
      header?: never
      path: {
        slug: string
      }
      cookie?: never
    }
    requestBody?: never
    responses: {
      /** @description OK */
      200: {
        headers: Record<string, unknown>
        content: {
          "application/json": {
            id: string
            /** @enum {string} */
            language: "id" | "en"
            title: string
            slug: string
            description: string | null
            /** @enum {string} */
            status: "published" | "draft" | "rejected" | "in_review"
            /** @enum {string} */
            visibility: "public" | "internal"
            metaTitle: string | null
            metaDescription: string | null
            topicTranslationId: string
            featuredImage: string | null
            createdAt: string | null
            updatedAt: string | null
          } | null
        }
      }
    }
  }
// ...
}